Convert to monorepo
