+++ examples/graphiql-example/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "graphiql-example",
+  "private": true,
+  "version": "1.0.0",
+  "dependencies": {
+    "apollo-server-express": "^1.3.6",
+    "body-parser": "^1.18.2",
+    "express": "^4.16.3",
+    "graphql": "^0.13.2",
+    "graphql-tools": "^3.0.0"
+  },
+  "devDependencies": {
+    "start-server-webpack-plugin": "^2.2.5",
+    "webpack": "^4.7.0",
+    "webpack-cli": "^2.1.3",
+    "webpack-node-externals": "^1.7.2"
+  },
+  "scripts": {
+    "start": "NODE_ENV=development webpack --watch"
+  }
+}
