+++ examples/graphiql-example/server/middleware/view.js
