+++ examples/graphiql-example/webpack.config.js
