+++ package.json
@@ -1,24 +1,14 @@
 {
-  "name": "graphql-mock-object",
-  "version": "1.0.0",
-  "main": "index.js",
   "author": "Eric Clemmons <eric@smarterspam.com>",
-  "license": "MIT",
-  "dependencies": {
-    "apollo-server-express": "^1.3.6",
-    "esm": "^3.0.28",
-    "express": "^4.16.3",
-    "graphql": "^0.13.2",
-    "graphql-tools": "^3.0.0",
-    "webpack": "^4.6.0",
-    "webpack-cli": "^2.0.15",
-    "webpack-node-externals": "^1.7.2"
-  },
-  "scripts": {
-    "dev": "NODE_ENV=development webpack --watch"
-  },
   "devDependencies": {
     "github-semantic-version": "^7.6.0",
-    "start-server-webpack-plugin": "^2.2.5"
-  }
+    "oao": "^1.5.1"
+  },
+  "name": "graphql-mock-object",
+  "private": true,
+  "version": "1.0.0",
+  "workspaces": [
+    "examples/*",
+    "packages/*"
+  ]
 }
