+++ packages/graphql-mock-object/index.js
@@ -1,0 +1,6 @@
+export { MockObject, resolvers } from "./resolvers"
+export { typeDefs } from "./typeDefs"
+
+export function mock() {
+  return {}
+}
