+++ packages/graphql-mock-object/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "graphql-mock-object",
+  "version": "1.0.0"
+}
